Remove AbstractProgressDialog, use concrete XbmcProgressDialog class instead
